requirements_service.txt
llama-index-embeddings-together
#llama-index-embeddings-voyageai
#llama-index-embeddings-openai
llama-index-embeddings-cohere
llama-index-embeddings-gemini
llama-index-embeddings-jinaai